requirements.txt
python-telegram-bot==20.7
selenium==4.15.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
resend==0.8.0
python-dotenv==1.0.0
aiosqlite==0.19.0
pillow==10.1.0
aiohttp==3.9.1
tenacity==8.2.3